+++ src/main/java/net/shasankp000/Entity/RespawnHandler.java
@@ -2,6 +2,7 @@ package net.shasankp000.Entity;
 
 import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.shasankp000.GameAI.BotEventHandler;
 
 public class RespawnHandler {
     public static void registerRespawnListener(ServerPlayerEntity bot) {
@@ -9,11 +10,11 @@ public class RespawnHandler {
         String botName = bot.getName().getString();
 
         ServerPlayerEvents.AFTER_RESPAWN.register((oldPlayer, newPlayer, alive) -> {
-            if (newPlayer.getName().getString().equals(botName)) { // Replace "Steve" with your bot's name
+            if (newPlayer.getName().getString().equals(botName)) {
                 System.out.println("Detected bot respawn for " + newPlayer.getName().getString());
-                AutoFaceEntity.handleBotRespawn(newPlayer); // Pass the new bot instance
+                BotEventHandler.onBotRespawn(newPlayer); // Ensure state reset and idle mode
+                AutoFaceEntity.handleBotRespawn(newPlayer); // Restart tick loop
             }
         });
     }
 }
-